avdovi/dovi_rpu{enc, dec}: fix ms_weight handling
The code as written was wrong. The l2.ms_weight value is coded as
a *signed* integer, rather than a shifted unsigned integer. (And even
so, the offset of 8192 would be too large, since 2^12 = 4096. Ditto for
l8.ms_weight, which should have an offset of 2048, not 8192)

In addition, because the l8.ms_weight struct member is unsigned, these
shifting semantics ended up overflowing the field, leading to undefined
behavior when transcoding. Fortunately, the damage was relatively
contained in practice, because it just corrupts the coding of this
field, which is ignored in practice in all implementations I have seen.

For some reason, Dolby decided to add a sign bit to l2.ms_weight but not
l3.ms_weight. In either case, it's quite probably that this is still
a shifted unsigned integer in disguise, since all samples seem to have
a value of 2048 for these fields. But without specs, that's guesswork,
and the official DV validator also reports these fields as unsigned
integers with a value of 2048.