avcodec/dovi_rpudec: replace brittle struct copying code
This code was unnecessarily trying to be robust against downgrades of
libavutil (relative to the version libavcodec was compiled against), but
in the process, ended up with very brittle code that is easy to
accidentally forget to update when adding new fields.

Instead, do the obvious thing and just directly copy the parts of the
struct known at compile time. Since it is not generally supported to
link against a version of libavutil older than the version libavcodec
was compiled against, the struct shrinking externally is not a case we
need to be worrying about.